@@ -1,8 +1,15 @@
 <script>
   import MainNavigation from '$lib/components/MainNavigation.svelte';
+  /**
+   * @typedef {Object} Props
+   * @property {import('svelte').Snippet} [children]
+   */
+
+  /** @type {Props} */
+  let { children } = $props();
 </script>
 
 <div class="my__layout">
   <MainNavigation />
-  <slot></slot>
+  {@render children?.()}
 </div>
